% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_covariate_metric}
\alias{compute_covariate_metric}
\title{Computes metric comparing covariates from two trees}
\usage{
compute_covariate_metric(rf, max_depth = NULL)
}
\arguments{
\item{rf}{randomForest object from which to compute similarities between trees}

\item{max_depth}{an option to set the maximum tree depth to consider when
comparing trees (set to NULL by default)}
}
\description{
Function for computing the fit metric from \insertCite{banerjee:2012;textual}{TreeTracer}.
}
\examples{

# Load packages
library(palmerpenguins)

# Load the Palmer penguins data
penguins <- na.omit(penguins)

# Fit a random forest
set.seed(71)
penguin_rf <-
  randomForest::randomForest(
    species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
    data = penguins,
    ntree = 5
  )

# Compute fit metrics between all trees
compute_covariate_metric(penguin_rf)
}
\references{
{
  \insertRef{banerjee:2012}{TreeTracer}
}
}
