% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fit_metric.R
\name{compute_fit_metric}
\alias{compute_fit_metric}
\title{Compute fit metric between two trees}
\usage{
compute_fit_metric(rf, data)
}
\arguments{
\item{rf}{randomForest object from which to compute similarities between trees}

\item{data}{data frame with predictor variables used to fit the model (does
not need to be the training data)}
}
\description{
The function computes the fit metric from \insertCite{chipman:1998;textual}{TreeTracer}
between two trees (classification or regression).
}
\examples{

# Load packages
library(palmerpenguins)

# Load the Palmer penguins data
penguins <- na.omit(penguins)

# Fit a random forest
set.seed(71)
penguin.rf <-
  randomForest::randomForest(
    species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
    data = penguins,
    ntree = 10
  )

# Compute fit metrics between all trees
compute_fit_metric(penguin.rf, penguins)
}
\references{
{
  \insertRef{chipman:1998}{TreeTracer}
}
}
