% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{compute_partition_metric}
\alias{compute_partition_metric}
\title{Computes partition metric between two trees}
\usage{
compute_partition_metric(rf, tree_preds)
}
\arguments{
\item{rf}{randomForest object from which to compute similarities between trees}

\item{tree_preds}{output from the function get_tree_preds}
}
\description{
Function for computing the partition metric from \insertCite{chipman:1998;textual}{TreeTracer}.
}
\examples{

# Load packages
library(palmerpenguins)

# Load the Palmer penguins data
penguins <- na.omit(penguins)

# Fit a random forest
set.seed(71)
penguin_rf <-
  randomForest::randomForest(
    species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
    data = penguins,
    ntree = 10
  )

# Predictions from all trees for five observations
tree_preds <- get_tree_preds(data = penguins[1:5,], rf = penguin_rf)

# Compute fit metrics between all trees
compute_partition_metric(penguin_rf, tree_preds)
}
\references{
{
  \insertRef{chipman:1998}{TreeTracer}
}
}
