% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_tree_preds.R
\name{get_all_tree_preds}
\alias{get_all_tree_preds}
\title{Get a prediction using a specified tree from a random forest}
\usage{
get_all_tree_preds(data, rf)
}
\arguments{
\item{data}{data frame with feature values for observations to obtain predictions}

\item{rf}{random forest model fit using randomForest}
}
\description{
Get a prediction using a specified tree from a random forest
}
\examples{

# Load the Palmer penguins data
library(palmerpenguins)
penguins <- na.omit(penguins)

# Fit a random forest
set.seed(71)
penguin_rf <-
  randomForest::randomForest(
    species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
    data = penguins
  )

# Extract tree data from the first tree in the random forest
get_tree_pred(penguins[1,], penguin_rf, 1)
}
