% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_matrix.R
\name{get_dist_matrix}
\alias{get_dist_matrix}
\title{Convert data frame of similarities from compute_fit_metric to distance matrix}
\usage{
get_dist_matrix(fit_metrics)
}
\arguments{
\item{fit_metrics}{output object from compute_fit_metric}
}
\description{
The function converts the data frame of similarities computed using
compute_fit_metric to a matrix of distances between all trees. Helpful to
use for clustering.
}
\examples{

# Load packages
library(palmerpenguins)

# Load the Palmer penguins data
penguins <- na.omit(penguins)

# Fit a random forest
set.seed(71)
penguin.rf <-
  randomForest::randomForest(
    species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
    data = penguins,
    ntree = 10
  )

# Compute fit metrics between all trees
fit_metrics <- compute_fit_metric(penguin.rf, penguins)

# Obtain the distance matrix
get_dist_matrix(fit_metrics)
}
