% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_depth.R
\name{get_node_depth}
\alias{get_node_depth}
\title{Determines the depth of each node in a tree}
\usage{
get_node_depth(df, node)
}
\arguments{
\item{df}{data frame of tree information generated with get_tree_data}

\item{node}{node number}
}
\description{
Determines the depth of each node in a tree
}
