% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trace_data.R
\name{get_trace_data}
\alias{get_trace_data}
\title{Convert tree information from a random forest to the structure for a trace plot}
\usage{
get_trace_data(tree_data, train, width = 0.8)
}
\arguments{
\item{tree_data}{data.frame obtained using get_tree_data}

\item{train}{features used to train the random forest which the tree is from}

\item{width}{specifies the width of the horizontal feature lines in a trace plot
(a number between 0 and 1; default is 0.8)}
}
\description{
Convert tree information from a random forest to the structure for a trace plot
}
\examples{

# Fit a random forest using the iris data
set.seed(71)
iris.rf <- randomForest::randomForest(Species ~ ., data = iris)

# Extract tree data from the first tree in the random forest
tree_data <- get_tree_data(iris.rf, 1)

# Obtain the trace data for the first tree in the random forest
get_trace_data(tree_data, iris[,-5])
}
