% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_data.R
\name{get_tree_data}
\alias{get_tree_data}
\title{Extract and organize information from a random forest tree}
\usage{
get_tree_data(rf, k)
}
\arguments{
\item{rf}{random forest model fit using randomForest}

\item{k}{number identifying the tree in the random forest from which
to extract information}
}
\description{
Extract and organize information from a random forest tree
}
\examples{

# Load the Palmer penguins data
penguins <- na.omit(palmerpenguins::penguins)

# Fit a random forest
set.seed(71)
penguin.rf <- randomForest::randomForest(species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g, data = penguins)

# Extract tree data from the first tree in the random forest
get_tree_data(penguin.rf, 1)
}
