% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_level.R
\name{get_tree_level}
\alias{get_tree_level}
\title{Determines the level of each node in a tree}
\usage{
get_tree_level(df, node)
}
\arguments{
\item{df}{data frame of tree information generated with get_tree_data}

\item{node}{node number}
}
\description{
Determines the level of each node in a tree
}
