% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_pred.R
\name{get_tree_pred}
\alias{get_tree_pred}
\title{Get a prediction using a specified tree from a random forest}
\usage{
get_tree_pred(obs, rf, k)
}
\arguments{
\item{obs}{data frame with feature values for an observation to get a prediction for}

\item{rf}{random forest model fit using randomForest}

\item{k}{number identifying the tree in the random forest from which
to extract information}
}
\description{
Get a prediction using a specified tree from a random forest
}
\examples{

# Load the Palmer penguins data
library(palmerpenguins)
penguins <- na.omit(penguins)

# Fit a random forest
set.seed(71)
penguin_rf <-
  randomForest::randomForest(
    species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
    data = penguins
  )

# Extract tree data from the first tree in the random forest
get_tree_pred(penguins[1,], penguin_rf, 1)
}
