% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_preds.R
\name{get_tree_preds}
\alias{get_tree_preds}
\title{Get predictions from the individual trees in a random forest}
\usage{
get_tree_preds(data, rf)
}
\arguments{
\item{data}{data frame with feature values for observations to obtain predictions}

\item{rf}{random forest model fit using randomForest}
}
\description{
Get predictions from the individual trees in a random forest
}
\examples{

# Load the Palmer penguins data
library(palmerpenguins)
penguins <- na.omit(penguins)

# Fit a random forest
set.seed(71)
penguin_rf <-
  randomForest::randomForest(
    species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
    data = penguins,
    ntree = 10
  )

# Extract tree data corresponding to the first five
# observations in the data
get_tree_preds(penguins[1,], penguin_rf)
}
