% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree_preds.R
\name{plot_tree_preds}
\alias{plot_tree_preds}
\title{Scatter plot matrix of predictions from two trees showing where they disagree}
\usage{
plot_tree_preds(rf, data, t1, t2)
}
\arguments{
\item{rf}{randomForest object from which to compute similarities between trees}

\item{data}{data frame with predictor variables used to fit the model (does
not need to be the training data)}

\item{t1}{tree ID for first tree in comparison}

\item{t2}{tree ID for second tree in comparison}
}
\description{
Given two tree IDs, the function creates a scatter plot matrix of the
predictor variables with the points colored by whether the predictions
from the two tree agree.
}
