% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_data.R
\name{get_tree_data}
\alias{get_tree_data}
\title{Extract and organize information from a random forest tree}
\usage{
get_tree_data(rf, k)
}
\arguments{
\item{rf}{random forest model fit using randomForest}

\item{k}{number identifying the tree in the random forest from which
to extract information}
}
\description{
Extract and organize information from a random forest tree
}
\examples{

# Fit a random forest using the iris data
set.seed(71)
iris.rf <- randomForest::randomForest(Species ~ ., data = iris)

# Extract tree data from the first tree in the random forest
get_tree_data(iris.rf, 1)
}
