% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Create a trace plot of trees from a random forest}
\usage{
trace_plot(
  rf,
  train,
  tree_ids,
  width = 0.8,
  alpha = 0.5,
  color_by_id = FALSE,
  facet_by_id = FALSE,
  nrow = NULL,
  max_depth = NULL,
  rep_tree = NULL,
  rep_tree_size = 1,
  rep_tree_color = "blue",
  rep_tree_alpha = 1
)
}
\arguments{
\item{rf}{random forest model fit using randomForest}

\item{train}{features used to train the random forest which the tree is from}

\item{tree_ids}{vector of numbers specifying the trees to include in the trace plot}

\item{width}{specifies the width of the horizontal feature lines in a trace plot
(a number between 0 and 1; default is 0.8)}

\item{alpha}{alpha value for the lines in the trace plot (a number between 0
and 1; default is 0.5)}

\item{color_by_id}{should the trace lines be colored by the tree IDs? (default if FALSE)}

\item{facet_by_id}{should the traces be faceted by tree IDs? (default if FALSE)}

\item{nrow}{number of rows if facet_by_id is TRUE (othewise ignored)}

\item{max_depth}{the deepest level to include in the trace plot (set to NULl by default)}

\item{rep_tree}{option to add a "representative tree" on top of the trace plot by providing
a data frame with the structure of the get_tree_data function (NULL by default)}

\item{rep_tree_size}{line size of "representative tree" (1 by default)}

\item{rep_tree_color}{line color of "representative tree" ("blue" by default)}

\item{rep_tree_alpha}{line alpha of "representative tree" (1 by default)}
}
\description{
Trace plots are useful tools for visually comparing trees from a random
forest. This functions creates a trace plot given a set of trees from a
random forest fit using the randomForest package. For more information on
trace plots, see \insertCite{urbanek:2008;textual}{TreeTracer}.
}
\examples{

# Load packages
library(dplyr)
library(palmerpenguins)

# Load the Palmer penguins data
penguins <- na.omit(penguins)

# Fit a random forest
set.seed(71)
penguin_rf <-
  randomForest::randomForest(
    species ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
    data = penguins
  )

# Generate a trace plot of the first 10 trees in the forest
trace_plot(
 rf = penguin_rf,
 train = penguins \%>\% select(bill_length_mm, bill_depth_mm, flipper_length_mm, body_mass_g),
 tree_ids = 1:10
)
}
\references{
{
  \insertRef{urbanek:2008}{TreeTracer}
}
}
