% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plot.R
\name{trace_plot}
\alias{trace_plot}
\title{Create a trace plot of trees from a random forest}
\usage{
trace_plot(rf, train, tree_ids, width = 0.8, alpha = 0.5)
}
\arguments{
\item{rf}{random forest model fit using randomForest}

\item{train}{features used to train the random forest which the tree is from}

\item{tree_ids}{vector of numbers specifying the trees to include in the trace plot}

\item{width}{specifies the width of the horizontal feature lines in a trace plot
(a number between 0 and 1; default is 0.8)}

\item{alpha}{alpha value for the lines in the trace plot (a number between 0 and 1)}
}
\description{
Trace plots are useful tools for visually comparing trees from a random
forest. This functions creates a trace plot given a set of trees from a
random forest fit using the randomForest package. For more information on
trace plots, see \insertCite{urbanek:2008;textual}{TreeTracer}.
}
\examples{

# Fit a random forest using the iris data
set.seed(71)
iris.rf <- randomForest::randomForest(Species ~ ., data = iris)

# Generate a trace plot of the first 25 trees in the forest
trace_plot(iris.rf, iris[,-5], 1:25)
}
\references{
{
  \insertRef{urbanek:2008}{TreeTracer}
}
}
